Hiding the part of the accordian item , and opening it only the item  which contains a special class of 'open'
to hide certain part we wrapped it insid a div and gave it a class 'hidden-box' and set the display property of this class as none.

and if we want to open certain item the when click, using JS we should add a another special class say 'open' to that item and set the property of this class such that it ervert back the display of the hidden-box to block instead of none.